Integrate citation folding with reftex
* czm-tex-fold.el (czm-tex-fold-bibtex-abbrev): Let-bind
case-fold-search to t (previously taken care of by caller).
(czm-tex-fold--find-bib-entry):
(czm-tex-fold--get-citation-info): New functions.
(czm-tex-fold-cite-display): Use them.
* README.org (Configuration): Update.

The idea is that now czm-tex-fold-bib-file acts as a "backup" option,
kicking in only if reftex can't find the entry for a given citation.
Currently this works only with bib files, not with thebibliography
sections.